TCD plugin: Tab-local Current working directory
This plugin provides the TCD command and SetTCD function. These two will
run :lcd for all windows in the current tab, thus making the whole tab
use a different working directory.

Opening new windows and tabs is no longer a problem as the "lcd" are
inherited by the new windows.