Let GVim itself deal with a files list argument.
It seems that lst2mlt and GVim automation will sometimes drop files,
maybe the automation calls come in too fast so that GVim can't keep up
loading. Simply using an expression which reads in the file list allows
GVim to open them all directly.

Only nitpick is that the "Tab-local Current Directory" is set for the
file list rather than the first file.