Initial version of ManicHelper
A script dedicated to make use of ManicTime easier.

It turns the numpad into a place for shortcuts which are normally
executed using the mouse. A couple of them are used to move the mouse
around (making this Yet Another Mouse-Move AHK Script), some temporarily
increase the mouse move speed, then there's a shortcut for Ctrl+Scroll
wheel to move the timeline around and two to zoom in or out.