Powershell: add GitReviewPR which grabs PR changes from Azure Devops
It will create a local tag review/PR<num> which points to a squashed
merge commit, which in turn based on the refs which Azure Devops exposes
for each pull request so that no merge magic needs to happen.
Once that is done it will also output a .review file with the changes
from the squashed merge in a sort of patch file.