Vim: add 'B' indicator after file encoding to show if file contains a BOM
While a Byte Order Mark must always be present in an UTF-16 file, it's
generally optional for UTF-8.
Until you encounter a tool which requires a 3-byte UTF-8 "BOM" sequence
to detect that the file is UTF-8 or crashes on this weird byte
sequence... Then you can better confirm if it's there.