Fix focus detection so that only main window keeps screenshots on top
It is not necessary to keep the screenshots window on top when creating a
tag, yet the focus detection would still keep showing it. It turns out
the focus detection matches on the window title and that the Create Tag
window has a similar title.

Changing the title match to require inclusion of a space character and a
hyphen narrows things down.