Searchengines: add custom OpenSearchDescription files and method to install
It wasn't easy to find a method which worked in Iridium, while Pale Moon
was way less picky.

The end result is a setup using ncat where it uses a little bit of Lua
script to act as a backend. This generates HTML which points to the OSD
files, then serves those with the right content-type headers.

No idea why Iridium wouldn't simply accept a new search engine defined
in any kind of HTML file, it seems it must be linked from the root of
the domain name. Lucky for us it doesn't mind query parameters, so the
parameter defines which engine is referenced (there can only be one at a
time).