TC: Remove ExtractHere, tweak built-in unpackage dialog, show root parent dir
Apparently there's a TC setting which will set the "Extract to new
folder, same as package name" flag by default, extracting into the
current folder is then possible by blanking the "Unpack files to" box
(which is easy to do because it is highlighted when the dialog opens).

Also enable the "Show parent '..' item in root dir'. While it doesn't do
anything most of the time, you _can_ right-click it to get a context
menu for the root of the drive.