Windows bin: add wrapper script to run git-cola in background
The git-cola authors didn't want to handle the multi-platform complexity
of launching git-cola in the background (i.e. making it so that running
it from a shell will not make that shell wait for git-cola to exit).

This wrapper script looks up where the real deal is further in the path,
then wrangles some paths to find the embedded pythonw instance in the
git-cola install dir.

Going through the PATH variable helps avoid an issue where git-cola is
installed in `~/AppData/Local` on one of my machines and in Program
Files on another.