TC: Run git cola through git
The real fix would be to write a custom statically compiled executable
which can launch git-cola in the background on any platform.

To be left for when I've got some more time, for now it has the drawback
that it opens a fresh console window which cannot be closed until
git-cola is done.