AHK: Add MouseLocker
First attempt at easing use of ManicTime, it would lock the mouse so
that you could move the mouse only sideways when travelling through the
timeline.

There's also a bonus mode which transports the cursor to the leftmost
screen and puts it right on top of the most likely place for a vertical
taskbar there. It then locks the mouse vertically so that one can pick
the task button to click.