Powershell: change GitKU to accept multiple branches and options
Arguments starting with - are treated as an argument and get passed
as-is. Everything else is taken for a branch, at which point GitKU will
look up the upstream branch and make sure to include it.

There's special support for HEAD so you don't have to type the current
branch name if you want to compare it to some other branch.

If no branch is specified then the current branch (HEAD) is still used
by default.