Vim: jump to line where the cursor was last time a file was open
Normally the default vimrc adds an autocmd to do that, since there's a
custom rc in this repo we need to add it once more.

Sourcing the default vimrc would work as well of course, it's just that
different versions of vim might introduce different defaults, so for now
we'll just stick to vim-sensible and whatever we add to the rc folder.