Less: align current line to center when searching
The --jump-target=.5 means that a search result will be shown in the
center of the screen.

It would be better if it worked more like Vim, but this setting is at
least better than not aligning anything so that a  search result might
end up as the first line of the screen (thus losing the context when
searching in e.g. a man page).