TC: Add git clean filter so config is not always marked as modified
TotalCommander overwrites a particular setting in its config file with a
unique value every time it starts to make sure that it can write and
read settings.

Then there are some history settings (like: "is quick search filter on")
which are remembered between restarts which are not worth pushing to the
dotfiles repo.

This git "clean" filter will replace those values with a fixed value,
thus avoiding marking the wincmd.ini file as changed when it's just one
of the expected values that flipped value.