Powershell: Compare-TTML makes Kdiff3 pre-process TTML files for aligning
Kdiff3 supports a pre-processor just after the files are read and one
between the moment lines are auto-aligned and actually compared for
changes.

Passing the -focusOnText flag will strip timing in that second phase,
which means that timing changes show up as "just whitespace changes",
but it _will_ improve the auto-aligment because that only has to work
with text.