Powershell: Compare-TTML ignore style attribute
Amara seems to like setting the paragraph style of each individual
paragraph.
SubtitleEdit doesn't, so just strip it while comparing.

Preprocessor was originally intended to contain a (x|y) regex but it
seems that Kdiff3 uses the | character as a command separator, in which
case it will run the left and right command in a pipe setup.