AHK: Disable next day / previous day shortcuts, focus detection
ManicTime now allows customization of a limited amount of shortcuts.
Previous and next day can be handled through the built-in features.

Also disable the focus detection-and-raise logic for the screenshots
window. It worked pretty decently, but it had the effect that the "Tag
time" menu would get partially obscured by the screenshot window (the
latter was constantly kept at a high Z-order). Using Windows 10's
multiple desktops and parking ManicTime on one on its own allows to keep
the ManicTime and Screenshots window side-to-side without any other
windows interfering.