Powershell: Script to convert a base64 string to Hex
Most tools decoding the clipboard try to render the decoded bytes as
UTF-8, but if the input contains non-displayable characters or e.g.
UTF-16 then it would be far more interesting to decode to a hex string.

This scripts allows one to inspect the base64-encoded contents on the
clipboard in a Hex editor without having to decode to write to a file
and decode that. Just pass a base64 string to the script and put the
result in the clipboard, then paste into a hex editor as a hex string.