Homemaker & TC: Replace Console-Z with Windows Terminal
It looks like Console-Z is languishing and Windows Terminal seems to
have equivalent features.

One of Console-Zs annoyances on modern Windows 10 is that locking the PC
would throw off window width and synchronization with its background
console (Console-Z does not use the ConPTY API yet) and needed to be
cleared to be readable. It also wouldn't handle MSys' copy of Vim very
well so it seems it is time to make the switch.