Homemaker & PS: Install and Import posh-git
It turns out that it is possible to stop posh-git from showing up in the
prompt (and thus having to check a repository's status all the time).

That means we can import it just for the command completers, which are
quite nice.

Use posh-git command completer for gitk, stop wrapping branch names in
single quotes when completing for custom git commands.