PS: Gitku command now uses ':' as special escape character
The '@' character is reserved in powershell so use ':' instead to avoid
having to quote things.

Also improve the command so that it can run with just a local branch if
one wants to, and select the first branch no matter if it is a local
branch or local/upstream pair.