PS: Function to hard reset branch to upstream
This function does a few checks, like making sure there is actually an
upstream, but the most important is that it bails when uncommitted
changes are present.

git reset --hard is needed to clean up the working copy in case of
switching to the upstream version of a branch, but it will happily blow
away uncommitted changes as well. This helper function prevents that.