Homemaker: let bootstrap script account for Zone.Identifier
When the dotfiles have been downloaded from github in the zip format for
bootstrapping, Windows might have placed the "download from the
Internet" zone identifier on the zip file.

If you then extract this using File Explorer, the zone identifier will
be copied to the extract files for protection. However, this will trip
up Powershell scripts as they are now untrusted without digital
signature.

This change will unblock everything (assuming the batch file was
unprotected first of course so that it can run).